Now using ESP8266WebServer
The ESP8266 now uses ESP8266WebServer to handle web requests. Using SPIFFS memory to store the web page. The ClientList class created on the previous commit is now useless :)

+ Json.get now returns a String containing the json instead of sending it to stream.
+ utils.clock now returns a String instead of using a char buffer.
+ Renamed:
    - ReqErr -> ErrorType
    - ReqMes -> MessageType
    - Req -> RequestData
+ Random optimizations
+ Secured my TimeZoneDB API key

This might actually be one of the biggest changes in the project in one commit in terms of features. There is now the ability to easily implement a fully functional graphical interface with no change to the code.