Fixed printing slowness with the SD card
Logging to the SD card was strangely slow when the card was loaded at boot time. This was due to the fact that the LED strip library messed up the SPI settings for the SD card. An easy fix was to simply initialize the SD card after the LED strip.

+ Cleaned Sd Card related code
+ Added the ability to make a new attempt when the loading of the card has failed
+ Removed the timer for flushing data

==================== Memory usage ====================
- Arduino:
    RAM:   [=======   ]  71.4% (used 5845 bytes from 8192 bytes)
    Flash: [==        ]  20.0% (used 50664 bytes from 253952 bytes)
- ESP8266:
    RAM:   [====      ]  36.9% (used 30248 bytes from 81920 bytes)
    Flash: [===       ]  33.1% (used 345840 bytes from 1044464 bytes)