Updated from SPIFFS to LittleFS
A big thanks to to StridingDragon and Yuriy for the solution (https://community.platformio.org/t/how-to-define-littlefs-partition-build-image-and-flash-it-on-esp32/11333/16)

SPIFFS has been deprecated and LittleFs is now the recommended way to go. The change is fairly simple, all calls to SPIFFS are replaced with call to LittleFs in the code and the upload script is changed to use mklittlefs (https://github.com/earlephilhower/mklittlefs/releases) thanks to a simple python script.
This trick shouldn't be necessary when LittleFs gets implemented in PlatformIO

==================== Memory usage ====================
- Arduino:
    RAM:   [=======   ]  68.0% (used 5571 bytes from 8192 bytes)
    Flash: [==        ]  20.7% (used 52510 bytes from 253952 bytes)
- ESP8266:
    RAM:   [====      ]  37.2% (used 30452 bytes from 81920 bytes)
    Flash: [===       ]  32.8% (used 342596 bytes from 1044464 bytes)