Changed Serial reception timeout
This is to avoid getting stuck for too long when receiving junk on the Serial input (When the ESP resets for example). The timeout is set according to whether the information comes from the debugSerial or the comSerial. The ESP8266 has the debugSerial timeout anyway because, for RX, debugSerial and comSerial are the same (This shouldn't be a problem as the Arduino doesn't send junk on its comSerial).

==================== Memory usage ====================
- Arduino:
    RAM:   [=======   ]  68.0% (used 5571 bytes from 8192 bytes)
    Flash: [==        ]  20.7% (used 52570 bytes from 253952 bytes)
- ESP8266:
    RAM:   [====      ]  37.1% (used 30420 bytes from 81920 bytes)
    Flash: [===       ]  32.8% (used 342580 bytes from 1044464 bytes)