Made light mode tasks creation more robust
Light mode tasks are now created by the active running task from a message sent to the queue.
The message gives the information of which mode to create or if the shutdown task should be created.
There is also now only one handle being stored for light mode task and shutdown task.

Signed-off-by: Farès Chati <faresticha@gmail.com>