Reduced responsability of json module to the minimum
The json module now uses callbacks and helper functions to generate a json document.

Any module that wants to contribute to a specific json document can "subscribe" to it and its callbacks functions will be called when necessary to generate its portion of the json.

Each subscription creates a JSON Object at the root of the json document after the Status and Message strings. The content of this Object is set from the callback function of the subscriber.

+ Fixed some LOG calls missing the TAG argument
+ Added an HTTP response in case of json failure
+ Fixed the full_line system for web color display
+ Deleted all light_get functions
+ Added some debug logging in light_set functions

Next step is to tackle the Request module.

========================== Memory usage ==========================
RAM:   [=         ]  14.3% (used 46732 bytes from 327680 bytes)
Flash: [=======   ]  67.1% (used 879545 bytes from 1310720 bytes)

Signed-off-by: Farès Chati <faresticha@gmail.com>